Move unit and ui tests dependencies to different modules (#2403)
