+++ examples/src/androidTest/java/com/mapbox/navigation/examples/activity/TripServiceActivityKtTest.kt
@@ -4,7 +4,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.uiautomator.By
 import androidx.test.uiautomator.Until
 import com.mapbox.navigation.examples.R
-import com.mapbox.navigation.testing.NotificationTestRule
+import com.mapbox.navigation.testing.ui.NotificationTestRule
 import com.schibsted.spain.barista.assertion.BaristaEnabledAssertions.assertEnabled
 import com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertContains
 import com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed
