+++ libtesting-ui/build.gradle
@@ -24,7 +24,8 @@ android {
 dependencies {
     api dependenciesList.androidxTestJunit
     api dependenciesList.junit
-    api dependenciesList.coroutinesTestAndroid
-    api dependenciesList.mockk
-    api dependenciesList.robolectric
+    api(dependenciesList.barista) {
+        exclude group: 'org.jetbrains.kotlin'
+        exclude group: 'androidx.annotation'
+    }
 }
