+++ libtesting-ui/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest xmlns:tools="http://schemas.android.com/tools"
+    package="com.mapbox.navigation.testing.ui">
+
+    <uses-sdk tools:overrideLibrary="com.schibsted.spain.barista, android_libs.ub_uiautomator" />
+</manifest>
