+++ settings.gradle
@@ -12,4 +12,5 @@ include ':app',
         ':liblogger',
         ':libnavigation-util',
         ':libnavigation-core',
+        ':libtesting-ui',
         ':libtesting'
